Add explicit casts in tests

+++ tests/expr7.py
@@ -1,10 +1,11 @@
+from ltypes import i32
 def test_pow():
     a: i32
-    a = pow(2, 2)
+    a = i32(pow(2, 2))
 
 def test_pow_1(a: i32, b: i32) -> i32:
     res: i32
-    res = pow(a, b)
+    res = i32(pow(a, b))
     return res
 
 def main0():
